Made sure Double types do not used the string based constructor for default values.
Preferring double to Biginteger since the backend datatype is python float which is same as kotlin double to avoid any conversion surprises.

+++ cli/src/main/kotlin/ai/divyam/cli/model/ModelPricingStore.kt
@@ -12,7 +12,6 @@ import java.io.File
 import java.io.FileInputStream
 import java.io.InputStream
 import java.time.LocalDate
-import java.math.BigDecimal
 
 @Reflectable
 data class ModelPricing(
@@ -21,10 +20,10 @@ data class ModelPricing(
     val currency: String,
 
     @param:JsonProperty("text_input_price")
-    val textInputPrice: BigDecimal,
+    val textInputPrice: Double,
 
     @param:JsonProperty("text_output_price")
-    val textOutputPrice: BigDecimal,
+    val textOutputPrice: Double,
 
     @param:JsonProperty("per_n_tokens")
     val perNTokens: Int,
