+++ cli/src/main/kotlin/ai/divyam/cli/sa/SaUpdateCommand.kt
@@ -10,7 +10,6 @@ import ai.divyam.data.model.ServiceAccountUpdateRequest
 import kotlinx.coroutines.runBlocking
 import picocli.CommandLine
 import picocli.CommandLine.Option
-import java.math.BigDecimal
 
 @CommandLine.Command(
     name = "update",
@@ -139,7 +138,7 @@ class SaUpdateCommand : BaseCommand(), HasSecurityPolicy {
                 trafficAllocationConfig = mapper.readValue(
                     it,
                     Map::class.java
-                ) as Map<String, BigDecimal>
+                ) as Map<String, Double>
             )
         }
